[AIRLFOW-YYY] Move sync_perm command to seperate file
